@@ -71,9 +71,10 @@ exclude = [
     ".git",
     ".venv",
     "__pycache__",
+    "spec",
     "venv",
 ]
-target-version = "py39"
+target-version = "py310"
 
 [tool.ruff.lint]
 select = [
